feat: cooldown between games

+++ include/net/message.h
@@ -7,7 +7,7 @@ typedef struct	s_msg
 {
 	char				msg[256];
 	struct sockaddr_in	sender;
-	float				time;
+	double				time;
 }	t_msg;
 
 #endif
